refactor(non breaking): additional code cleanup
